requirements.txt
fastapi
uvicorn
line-bot-sdk
stripe
python-dotenv
google-genai
